pyproject.toml
[tool.poetry]
name = "gpt-assistants-api-ui"
version = "0.1.0"
description = ""
authors = ["ryo-ma <saka_ro@yahoo.co.jp>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10 || ^3.11"
streamlit = "^1.28.2"
openai = "^1.3.2"
black = "^23.12.1"
python-dotenv = "^1.0.1"
streamlit-authenticator = "^0.3.2"
PyMuPDF = "^1.25.5"
python-docx = "^0.8.11"
beautifulsoup4 = "^4.11.2"
mammoth = "^1.5.1"

[tool.poetry.group.develop.dependencies]
black = "^23.11.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
